Move the logic for form array in form elements
